my-website/backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic-settings==2.1.0

# LLM & Embeddings
google-generativeai==0.3.2

# Vector Database
qdrant-client==1.7.3

# Database
asyncpg==0.29.0

# PDF Processing
PyMuPDF==1.23.26

# Utilities
python-multipart==0.0.6
aiofiles==23.2.1
